Refactor sensor handling
Refactor away JoyAxis objects contained in JoySensors and the helper
sensor axes in SetJoystick. Move all used event handling logic into
JoySensor and remove dead sensor code from JoyAxis.
Use JoySensor::Type instead of int refer to a sensor
in SetJoystick. Convert all sensor values as floats as this is the
format SDL uses as well.

In joystickstatuswindow, scale all values by 1000 because QProgressBar
only supports integer values. Add some heler functions here to handle
the new sensor "moved" event with 3 float arguments to update the
progress bars accordingly.

+++ src/gui/joysensoreditdialog.h
@@ -49,7 +49,7 @@ class JoySensorEditDialog : public QDialog
   private slots:
     void implementPresets(int index);
 
-    void refreshSensorStats(int x, int y, int z);
+    void refreshSensorStats(float x, float y, float z);
 
     void checkMaxZone(int value);
     void openMouseSettingsDialog();
